+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Dark Theme: The search/filter result list has explicit colors: they are now readable (#1629)
 - UniversalMusicScraper: If an artist has no Discogs link on MusicBrainz, MediaElch was stuck.
 - ADE: Overview and posters are properly scraped again (#1650)
+- IMDb: Fix scraping of episode IDs from season pages.
 
 ### Changed